refactor(host-application): split runtime startup orchestration
Extract startup status, single-flight coordination, and workspace runtime startup into focused modules while keeping AppService as the public facade. This preserves runtime behavior while making the runtime start path easier to reason about and validate.
