+++ apps/desktop/src-tauri/crates/host-application/src/app_service/runtime_orchestrator.rs
@@ -1,72 +1,18 @@
+mod ensure_flight;
 mod registry;
 mod repo_health;
 mod repo_health_snapshot;
+mod start_pipeline;
 mod startup;
+mod startup_status;
+mod workspace_runtime;
 
 use super::AppService;
 use anyhow::{anyhow, Result};
 use host_domain::{
-    now_rfc3339, AgentRuntimeKind, RepoRuntimeStartupFailureKind, RepoRuntimeStartupStage,
-    RepoRuntimeStartupStatus, RunState, RunSummary, RuntimeDescriptor, RuntimeInstanceSummary,
-    RuntimeRole, RuntimeRoute,
+    AgentRuntimeKind, RunState, RunSummary, RuntimeDescriptor, RuntimeInstanceSummary, RuntimeRole,
 };
 use std::collections::{HashMap, HashSet};
-use std::process::Child;
-use std::sync::Arc;
-use std::time::Instant;
-
-#[derive(Clone, Copy)]
-pub(super) struct RuntimeExistingLookup<'a> {
-    repo_key: &'a str,
-    role: RuntimeRole,
-    task_id: Option<&'a str>,
-}
-
-pub(crate) struct RuntimePostStartPolicy<'a> {
-    existing_lookup: RuntimeExistingLookup<'a>,
-    prune_error_context: String,
-}
-
-pub(crate) struct RuntimeStartInput<'a> {
-    pub(crate) runtime_kind: AgentRuntimeKind,
-    pub(crate) startup_scope: &'a str,
-    pub(crate) repo_path: &'a str,
-    pub(crate) repo_key: String,
-    pub(crate) startup_started_at_instant: Instant,
-    pub(crate) startup_started_at: String,
-    pub(crate) task_id: &'a str,
-    pub(crate) role: RuntimeRole,
-    pub(crate) startup_policy: super::RuntimeStartupReadinessPolicy,
-    pub(crate) working_directory: String,
-    pub(crate) cleanup_target: Option<super::RuntimeCleanupTarget>,
-    pub(crate) tracking_error_context: &'static str,
-    pub(crate) startup_error_context: String,
-    pub(crate) post_start_policy: Option<RuntimePostStartPolicy<'a>>,
-}
-
-#[derive(Clone)]
-pub(in crate::app_service) struct RuntimeStartupProgress {
-    pub(in crate::app_service) started_at_instant: Instant,
-    pub(in crate::app_service) started_at: String,
-    pub(in crate::app_service) attempts: Option<u32>,
-    pub(in crate::app_service) elapsed_ms: Option<u64>,
-}
-
-struct RuntimeStartupFailure {
-    failure_kind: RepoRuntimeStartupFailureKind,
-    failure_reason: String,
-    detail: String,
-}
-
-pub(super) struct SpawnedRuntimeServer {
-    runtime_id: String,
-    runtime_route: RuntimeRoute,
-    child: Option<Child>,
-    runtime_process_guard: Option<super::RuntimeProcessGuard>,
-    startup_started_at_instant: Instant,
-    startup_started_at: String,
-    startup_report: super::RuntimeStartupWaitReport,
-}
 
 #[derive(Clone)]
 struct RunExposureCandidate {
@@ -147,349 +93,7 @@ impl RunExposurePlan {
     }
 }
 
-struct RuntimeEnsureFlightGuard<'a> {
-    service: &'a AppService,
-    runtime_kind: AgentRuntimeKind,
-    repo_key: String,
-    flight: Arc<super::service_core::RuntimeEnsureFlight>,
-    completed: bool,
-}
-
-impl<'a> RuntimeEnsureFlightGuard<'a> {
-    fn new(
-        service: &'a AppService,
-        runtime_kind: AgentRuntimeKind,
-        repo_key: &str,
-        flight: Arc<super::service_core::RuntimeEnsureFlight>,
-    ) -> Self {
-        Self {
-            service,
-            runtime_kind,
-            repo_key: repo_key.to_string(),
-            flight,
-            completed: false,
-        }
-    }
-
-    fn complete(&mut self, result: &Result<RuntimeInstanceSummary>) -> Result<()> {
-        self.completed = true;
-        self.service.complete_runtime_ensure_flight(
-            self.runtime_kind.clone(),
-            self.repo_key.as_str(),
-            &self.flight,
-            result,
-        )
-    }
-}
-
-impl Drop for RuntimeEnsureFlightGuard<'_> {
-    fn drop(&mut self) {
-        if self.completed {
-            return;
-        }
-
-        let aborted = Err(anyhow!("Runtime ensure aborted unexpectedly"));
-        if let Err(error) = self.service.complete_runtime_ensure_flight(
-            self.runtime_kind.clone(),
-            self.repo_key.as_str(),
-            &self.flight,
-            &aborted,
-        ) {
-            eprintln!(
-                "OpenDucktor warning: failed completing runtime ensure flight after abort: {error:#}"
-            );
-        }
-    }
-}
-
 impl AppService {
-    fn runtime_ensure_flight_key(runtime_kind: &AgentRuntimeKind, repo_key: &str) -> String {
-        format!("{}::{repo_key}", runtime_kind.as_str())
-    }
-
-    fn update_runtime_startup_status(
-        &self,
-        runtime_kind: &AgentRuntimeKind,
-        repo_key: &str,
-        update: impl FnOnce(&mut super::service_core::RuntimeStartupStatusEntry),
-    ) -> Result<()> {
-        let key = Self::runtime_ensure_flight_key(runtime_kind, repo_key);
-        let mut statuses = self
-            .runtime_startup_status
-            .lock()
-            .map_err(|_| anyhow!("Runtime startup status lock poisoned"))?;
-        let entry = statuses.entry(key).or_insert_with(|| {
-            super::service_core::RuntimeStartupStatusEntry::new(
-                runtime_kind.clone(),
-                repo_key.to_string(),
-                RepoRuntimeStartupStage::Idle,
-            )
-        });
-        update(entry);
-        entry.updated_at = now_rfc3339();
-        Ok(())
-    }
-
-    pub(in crate::app_service) fn mark_runtime_startup_requested(
-        &self,
-        runtime_kind: &AgentRuntimeKind,
-        repo_key: &str,
-        progress: &RuntimeStartupProgress,
-    ) -> Result<()> {
-        self.update_runtime_startup_status(runtime_kind, repo_key, |entry| {
-            entry.stage = RepoRuntimeStartupStage::StartupRequested;
-            entry.runtime = None;
-            entry.started_at = Some(progress.started_at.clone());
-            entry.started_at_instant = Some(progress.started_at_instant);
-            entry.elapsed_ms = None;
-            entry.attempts = Some(0);
-            entry.failure_kind = None;
-            entry.failure_reason = None;
-            entry.detail = None;
-        })
-    }
-
-    pub(in crate::app_service) fn mark_runtime_startup_waiting(
-        &self,
-        runtime_kind: &AgentRuntimeKind,
-        repo_key: &str,
-        progress: &RuntimeStartupProgress,
-    ) -> Result<()> {
-        self.update_runtime_startup_status(runtime_kind, repo_key, |entry| {
-            entry.stage = RepoRuntimeStartupStage::WaitingForRuntime;
-            entry.started_at = Some(progress.started_at.clone());
-            entry.started_at_instant = Some(progress.started_at_instant);
-            entry.elapsed_ms = None;
-            entry.attempts = progress.attempts;
-        })
-    }
-
-    pub(in crate::app_service) fn mark_runtime_startup_ready(
-        &self,
-        runtime_kind: &AgentRuntimeKind,
-        repo_key: &str,
-        runtime: &RuntimeInstanceSummary,
-        progress: &RuntimeStartupProgress,
-    ) -> Result<()> {
-        self.update_runtime_startup_status(runtime_kind, repo_key, |entry| {
-            entry.stage = RepoRuntimeStartupStage::RuntimeReady;
-            entry.runtime = Some(runtime.clone());
-            entry.started_at = Some(progress.started_at.clone());
-            entry.started_at_instant = Some(progress.started_at_instant);
-            entry.elapsed_ms = progress.elapsed_ms;
-            entry.attempts = progress.attempts;
-            entry.failure_kind = None;
-            entry.failure_reason = None;
-            entry.detail = None;
-        })
-    }
-
-    fn mark_runtime_startup_failed(
-        &self,
-        runtime_kind: &AgentRuntimeKind,
-        repo_key: &str,
-        progress: &RuntimeStartupProgress,
-        failure: RuntimeStartupFailure,
-    ) -> Result<()> {
-        self.update_runtime_startup_status(runtime_kind, repo_key, |entry| {
-            entry.stage = RepoRuntimeStartupStage::StartupFailed;
-            entry.runtime = None;
-            entry.started_at = Some(progress.started_at.clone());
-            entry.started_at_instant = Some(progress.started_at_instant);
-            entry.elapsed_ms = progress.elapsed_ms;
-            entry.attempts = progress.attempts;
-            entry.failure_kind = Some(failure.failure_kind);
-            entry.failure_reason = Some(failure.failure_reason.clone());
-            entry.detail = Some(failure.detail.clone());
-        })
-    }
-
-    fn clear_runtime_startup_status(
-        &self,
-        runtime_kind: &AgentRuntimeKind,
-        repo_key: &str,
-    ) -> Result<()> {
-        let key = Self::runtime_ensure_flight_key(runtime_kind, repo_key);
-        let mut statuses = self
-            .runtime_startup_status
-            .lock()
-            .map_err(|_| anyhow!("Runtime startup status lock poisoned"))?;
-        statuses.remove(key.as_str());
-        Ok(())
-    }
-
-    fn clear_runtime_startup_status_for_runtime(
-        &self,
-        runtime: &RuntimeInstanceSummary,
-    ) -> Result<()> {
-        self.clear_runtime_startup_status(&runtime.kind, runtime.repo_path.as_str())
-    }
-
-    pub fn runtime_startup_status(
-        &self,
-        runtime_kind: &str,
-        repo_path: &str,
-    ) -> Result<RepoRuntimeStartupStatus> {
-        let runtime_kind = self.resolve_supported_runtime_kind(runtime_kind)?;
-        let repo_key = self.resolve_authorized_repo_path(repo_path)?;
-        let status_key = Self::runtime_ensure_flight_key(&runtime_kind, repo_key.as_str());
-
-        if let Some(snapshot) = self
-            .runtime_startup_status
-            .lock()
-            .map_err(|_| anyhow!("Runtime startup status lock poisoned"))?
-            .get(status_key.as_str())
-            .cloned()
-        {
-            if snapshot.stage != RepoRuntimeStartupStage::RuntimeReady
-                || self
-                    .find_existing_workspace_runtime(&runtime_kind, repo_key.as_str())?
-                    .is_some()
-            {
-                return Ok(snapshot.to_public_status());
-            }
-        }
-
-        if let Some(runtime) =
-            self.find_existing_workspace_runtime(&runtime_kind, repo_key.as_str())?
-        {
-            return Ok(RepoRuntimeStartupStatus {
-                runtime_kind,
-                repo_path: repo_key,
-                stage: RepoRuntimeStartupStage::RuntimeReady,
-                runtime: Some(runtime.clone()),
-                started_at: Some(runtime.started_at.clone()),
-                updated_at: runtime.started_at,
-                elapsed_ms: None,
-                attempts: None,
-                failure_kind: None,
-                failure_reason: None,
-                detail: None,
-            });
-        }
-
-        Ok(RepoRuntimeStartupStatus {
-            runtime_kind,
-            repo_path: repo_key,
-            stage: RepoRuntimeStartupStage::Idle,
-            runtime: None,
-            started_at: None,
-            updated_at: now_rfc3339(),
-            elapsed_ms: None,
-            attempts: None,
-            failure_kind: None,
-            failure_reason: None,
-            detail: None,
-        })
-    }
-
-    fn acquire_runtime_ensure_flight(
-        &self,
-        runtime_kind: AgentRuntimeKind,
-        repo_key: &str,
-    ) -> Result<(Arc<super::service_core::RuntimeEnsureFlight>, bool)> {
-        let key = Self::runtime_ensure_flight_key(&runtime_kind, repo_key);
-        let mut flights = self
-            .runtime_ensure_flights
-            .lock()
-            .map_err(|_| anyhow!("Runtime ensure coordination state lock poisoned"))?;
-        if let Some(existing) = flights.get(key.as_str()) {
-            return Ok((existing.clone(), false));
-        }
-
-        let flight = Arc::new(super::service_core::RuntimeEnsureFlight::new());
-        flights.insert(key, flight.clone());
-        Ok((flight, true))
-    }
-
-    fn complete_runtime_ensure_flight(
-        &self,
-        runtime_kind: AgentRuntimeKind,
-        repo_key: &str,
-        flight: &Arc<super::service_core::RuntimeEnsureFlight>,
-        result: &Result<RuntimeInstanceSummary>,
-    ) -> Result<()> {
-        let stored_result = match result {
-            Ok(summary) => Ok(summary.clone()),
-            Err(error) => Err(format!("{error:#}")),
-        };
-        let mut poisoned = false;
-
-        {
-            let mut state = match flight.state.lock() {
-                Ok(state) => state,
-                Err(poisoned_state) => {
-                    poisoned = true;
-                    poisoned_state.into_inner()
-                }
-            };
-            *state =
-                super::service_core::RuntimeEnsureFlightState::Finished(Box::new(stored_result));
-            flight.condvar.notify_all();
-        }
-
-        {
-            let mut flights = match self.runtime_ensure_flights.lock() {
-                Ok(flights) => flights,
-                Err(poisoned_flights) => {
-                    poisoned = true;
-                    poisoned_flights.into_inner()
-                }
-            };
-            let key = Self::runtime_ensure_flight_key(&runtime_kind, repo_key);
-            flights.remove(key.as_str());
-        }
-
-        if poisoned {
-            return Err(anyhow!("Runtime ensure coordination state lock poisoned"));
-        }
-
-        Ok(())
-    }
-
-    fn wait_for_runtime_ensure_flight(
-        flight: &Arc<super::service_core::RuntimeEnsureFlight>,
-    ) -> Result<RuntimeInstanceSummary> {
-        let mut state = flight
-            .state
-            .lock()
-            .map_err(|_| anyhow!("Runtime ensure coordination state lock poisoned"))?;
-        loop {
-            match &*state {
-                super::service_core::RuntimeEnsureFlightState::Starting => {
-                    state = flight
-                        .condvar
-                        .wait(state)
-                        .map_err(|_| anyhow!("Runtime ensure coordination state lock poisoned"))?;
-                }
-                super::service_core::RuntimeEnsureFlightState::Finished(result) => {
-                    return result.as_ref().clone().map_err(|message| anyhow!(message));
-                }
-            }
-        }
-    }
-
-    fn find_existing_workspace_runtime(
-        &self,
-        runtime_kind: &AgentRuntimeKind,
-        repo_key: &str,
-    ) -> Result<Option<RuntimeInstanceSummary>> {
-        let mut runtimes = self
-            .agent_runtimes
-            .lock()
-            .map_err(|_| anyhow!("Agent runtime state lock poisoned"))?;
-        self.prune_stale_runtimes(&mut runtimes)?;
-        Ok(Self::find_existing_runtime(
-            &runtimes,
-            RuntimeExistingLookup {
-                repo_key,
-                role: Self::WORKSPACE_RUNTIME_ROLE,
-                task_id: None,
-            },
-        )
-        .filter(|runtime| runtime.kind == *runtime_kind))
-    }
-
     pub(super) fn ensure_runtime_supports_all_workflow_scopes(
         &self,
         runtime_kind: AgentRuntimeKind,
@@ -672,138 +276,6 @@ impl AppService {
         Ok(list)
     }
 
-    fn ensure_workspace_runtime(
-        &self,
-        runtime_kind: AgentRuntimeKind,
-        repo_path: &str,
-    ) -> Result<RuntimeInstanceSummary> {
-        let repo_key = self.resolve_authorized_repo_path(repo_path)?;
-        let repo_path = repo_key.as_str();
-
-        if let Some(existing) =
-            self.find_existing_workspace_runtime(&runtime_kind, repo_key.as_str())?
-        {
-            return Ok(existing);
-        }
-
-        let (flight, is_leader) =
-            self.acquire_runtime_ensure_flight(runtime_kind.clone(), repo_key.as_str())?;
-        if !is_leader {
-            return Self::wait_for_runtime_ensure_flight(&flight);
-        }
-        let mut flight_guard =
-            RuntimeEnsureFlightGuard::new(self, runtime_kind.clone(), repo_key.as_str(), flight);
-        let startup_started_at_instant = Instant::now();
-        let startup_started_at = now_rfc3339();
-
-        let startup_result = (|| -> Result<RuntimeInstanceSummary> {
-            if let Some(existing) =
-                self.find_existing_workspace_runtime(&runtime_kind, repo_key.as_str())?
-            {
-                return Ok(existing);
-            }
-
-            let startup_error_context = format!(
-                "{} workspace runtime failed to start for {repo_path}",
-                runtime_kind.as_str()
-            );
-            let startup_policy = self.resolve_runtime_startup_policy(
-                &runtime_kind,
-                "workspace_runtime",
-                repo_path,
-                Self::WORKSPACE_RUNTIME_TASK_ID,
-                Self::WORKSPACE_RUNTIME_ROLE,
-                startup_error_context.as_str(),
-            )?;
-
-            self.spawn_and_register_runtime(RuntimeStartInput {
-                runtime_kind: runtime_kind.clone(),
-                startup_scope: "workspace_runtime",
-                repo_path,
-                repo_key: repo_key.clone(),
-                startup_started_at_instant,
-                startup_started_at: startup_started_at.clone(),
-                task_id: Self::WORKSPACE_RUNTIME_TASK_ID,
-                role: Self::WORKSPACE_RUNTIME_ROLE,
-                startup_policy,
-                working_directory: repo_key.clone(),
-                cleanup_target: None,
-                tracking_error_context: "Failed tracking spawned OpenCode workspace runtime",
-                startup_error_context,
-                post_start_policy: Some(RuntimePostStartPolicy {
-                    existing_lookup: RuntimeExistingLookup {
-                        repo_key: repo_key.as_str(),
-                        role: Self::WORKSPACE_RUNTIME_ROLE,
-                        task_id: None,
-                    },
-                    prune_error_context: format!(
-                        "Failed pruning stale runtimes while finalizing workspace runtime for {repo_path}"
-                    ),
-                }),
-            })
-        })();
-        if let Err(error) = startup_result.as_ref() {
-            let startup_failure = error
-                .chain()
-                .find_map(|cause| cause.downcast_ref::<super::OpencodeStartupWaitFailure>());
-            let (failure_kind, failure_reason, attempts, elapsed_ms) = match startup_failure {
-                Some(failure) => (
-                    if failure.reason == "timeout" {
-                        RepoRuntimeStartupFailureKind::Timeout
-                    } else {
-                        RepoRuntimeStartupFailureKind::Error
-                    },
-                    failure.reason,
-                    Some(failure.report().attempts()),
-                    Some(failure.report().startup_ms()),
-                ),
-                None => (RepoRuntimeStartupFailureKind::Error, "error", None, None),
-            };
-            self.mark_runtime_startup_failed(
-                &runtime_kind,
-                repo_key.as_str(),
-                &RuntimeStartupProgress {
-                    started_at_instant: startup_started_at_instant,
-                    started_at: startup_started_at.clone(),
-                    attempts,
-                    elapsed_ms,
-                },
-                RuntimeStartupFailure {
-                    failure_kind,
-                    failure_reason: failure_reason.to_string(),
-                    detail: format!("{error:#}"),
-                },
-            )?;
-        }
-        flight_guard.complete(&startup_result)?;
-        startup_result
-    }
-
-    fn spawn_and_register_runtime(
-        &self,
-        input: RuntimeStartInput<'_>,
-    ) -> Result<RuntimeInstanceSummary> {
-        let spawned_server = self.spawn_runtime_server(&input)?;
-        let startup_started_at_instant = spawned_server.startup_started_at_instant;
-        let startup_started_at = spawned_server.startup_started_at.clone();
-        let startup_report = spawned_server.startup_report;
-        let runtime_kind = input.runtime_kind.clone();
-        let repo_key = input.repo_key.clone();
-        let summary = self.attach_runtime_session(input, spawned_server)?;
-        self.mark_runtime_startup_ready(
-            &runtime_kind,
-            repo_key.as_str(),
-            &summary,
-            &RuntimeStartupProgress {
-                started_at_instant: startup_started_at_instant,
-                started_at: startup_started_at,
-                attempts: Some(startup_report.attempts()),
-                elapsed_ms: Some(startup_report.startup_ms()),
-            },
-        )?;
-        Ok(summary)
-    }
-
     pub(super) fn resolve_supported_runtime_kind(
         &self,
         runtime_kind: &str,
@@ -837,26 +309,22 @@ fn collect_build_external_session_ids_for_run(
 
 #[cfg(test)]
 mod tests {
-    use super::{
-        AppService, RuntimeEnsureFlightGuard, RuntimeStartupFailure, RuntimeStartupProgress,
-    };
+    use super::AppService;
     use crate::app_service::test_support::{
         build_service_with_state, builtin_opencode_runtime_descriptor, make_task,
     };
     use crate::app_service::{AgentRuntimeProcess, RunProcess};
-    use anyhow::{anyhow, Result};
+    use anyhow::Result;
     use chrono::{TimeDelta, Utc};
     use host_domain::{
         now_rfc3339, AgentRuntimeKind, AgentSessionDocument, RepoRuntimeHealthMcp,
         RepoRuntimeHealthObservation, RepoRuntimeHealthRuntime, RepoRuntimeHealthState,
-        RepoRuntimeMcpStatus, RepoRuntimeStartupFailureKind, RepoRuntimeStartupStage, RunSummary,
-        RuntimeRoute, TaskStatus,
+        RepoRuntimeMcpStatus, RunSummary, RuntimeRoute, TaskStatus,
     };
     use host_infra_system::RepoConfig;
     use std::io::{Read, Write};
     use std::net::TcpListener;
     use std::process::Command;
-    use std::thread;
     use std::time::{Duration, Instant};
 
     fn spawn_opencode_session_status_server(
@@ -996,186 +464,6 @@ mod tests {
                     cleanup_target: None,
                 },
             );
-        Ok(())
-    }
-
-    #[test]
-    fn runtime_ensure_flight_guard_finishes_waiters_when_dropped_uncompleted() -> Result<()> {
-        let (service, _task_state, _git_state) = build_service_with_state(vec![]);
-        let repo_key = "/tmp/runtime-flight-guard";
-        let (flight, is_leader) =
-            service.acquire_runtime_ensure_flight(AgentRuntimeKind::opencode(), repo_key)?;
-        assert!(is_leader);
-
-        {
-            let _guard = RuntimeEnsureFlightGuard::new(
-                &service,
-                AgentRuntimeKind::opencode(),
-                repo_key,
-                flight.clone(),
-            );
-        }
-
-        let error = AppService::wait_for_runtime_ensure_flight(&flight)
-            .expect_err("dropped leader should finish waiters with an error");
-        assert!(error
-            .to_string()
-            .contains("Runtime ensure aborted unexpectedly"));
-
-        Ok(())
-    }
-
-    #[test]
-    fn complete_runtime_ensure_flight_recovers_poisoned_state_and_removes_entry() -> Result<()> {
-        let (service, _task_state, _git_state) = build_service_with_state(vec![]);
-        let repo_key = "/tmp/runtime-flight-poison";
-        let (flight, is_leader) =
-            service.acquire_runtime_ensure_flight(AgentRuntimeKind::opencode(), repo_key)?;
-        assert!(is_leader);
-
-        let poison_handle = thread::spawn({
-            let flight = flight.clone();
-            move || {
-                let _lock = flight
-                    .state
-                    .lock()
-                    .expect("flight state should be available for poisoning");
-                panic!("poison runtime ensure flight state");
-            }
-        });
-        assert!(poison_handle.join().is_err());
-
-        let error = service
-            .complete_runtime_ensure_flight(
-                AgentRuntimeKind::opencode(),
-                repo_key,
-                &flight,
-                &Err(anyhow!("simulated startup failure")),
-            )
-            .expect_err("poisoned completion should surface an error");
-        assert!(error
-            .to_string()
-            .contains("Runtime ensure coordination state lock poisoned"));
-
-        let flights = service
-            .runtime_ensure_flights
-            .lock()
-            .expect("runtime ensure flights lock should remain available");
-        assert!(flights.is_empty());
-
-        Ok(())
-    }
-
-    #[test]
-    fn runtime_startup_status_tracks_waiting_and_failure_stages() -> Result<()> {
-        let (service, _task_state, _git_state) = build_service_with_state(vec![]);
-        let repo_path = "/tmp/runtime-startup-status";
-        let started_at_instant = Instant::now();
-        let started_at = "2026-04-04T16:00:00Z";
-
-        service.mark_runtime_startup_requested(
-            &AgentRuntimeKind::opencode(),
-            repo_path,
-            &RuntimeStartupProgress {
-                started_at_instant,
-                started_at: started_at.to_string(),
-                attempts: Some(0),
-                elapsed_ms: None,
-            },
-        )?;
-        service.mark_runtime_startup_waiting(
-            &AgentRuntimeKind::opencode(),
-            repo_path,
-            &RuntimeStartupProgress {
-                started_at_instant,
-                started_at: started_at.to_string(),
-                attempts: Some(3),
-                elapsed_ms: None,
-            },
-        )?;
-
-        let waiting_status = service.runtime_startup_status("opencode", repo_path)?;
-        assert_eq!(
-            waiting_status.stage,
-            RepoRuntimeStartupStage::WaitingForRuntime
-        );
-        assert_eq!(waiting_status.attempts, Some(3));
-        assert_eq!(waiting_status.started_at.as_deref(), Some(started_at));
-
-        service.mark_runtime_startup_failed(
-            &AgentRuntimeKind::opencode(),
-            repo_path,
-            &RuntimeStartupProgress {
-                started_at_instant,
-                started_at: started_at.to_string(),
-                attempts: Some(4),
-                elapsed_ms: Some(4200),
-            },
-            RuntimeStartupFailure {
-                failure_kind: RepoRuntimeStartupFailureKind::Timeout,
-                failure_reason: "timeout".to_string(),
-                detail: "OpenCode startup probe failed reason=timeout".to_string(),
-            },
-        )?;
-
-        let failed_status = service.runtime_startup_status("opencode", repo_path)?;
-        assert_eq!(failed_status.stage, RepoRuntimeStartupStage::StartupFailed);
-        assert_eq!(
-            failed_status.failure_kind,
-            Some(RepoRuntimeStartupFailureKind::Timeout)
-        );
-        assert_eq!(failed_status.failure_reason.as_deref(), Some("timeout"));
-        assert_eq!(failed_status.attempts, Some(4));
-        assert_eq!(failed_status.elapsed_ms, Some(4200));
-
-        Ok(())
-    }
-
-    #[test]
-    fn runtime_startup_status_ignores_stale_ready_snapshot_without_live_runtime() -> Result<()> {
-        let (service, _task_state, _git_state) = build_service_with_state(vec![]);
-        let repo_path = "/tmp/runtime-stale-ready";
-        let started_at_instant = Instant::now();
-        let started_at = "2026-04-04T16:00:00Z";
-
-        service.mark_runtime_startup_waiting(
-            &AgentRuntimeKind::opencode(),
-            repo_path,
-            &RuntimeStartupProgress {
-                started_at_instant,
-                started_at: started_at.to_string(),
-                attempts: Some(1),
-                elapsed_ms: None,
-            },
-        )?;
-        service.mark_runtime_startup_ready(
-            &AgentRuntimeKind::opencode(),
-            repo_path,
-            &host_domain::RuntimeInstanceSummary {
-                kind: AgentRuntimeKind::opencode(),
-                runtime_id: "runtime-stale".to_string(),
-                repo_path: repo_path.to_string(),
-                task_id: None,
-                role: host_domain::RuntimeRole::Workspace,
-                working_directory: repo_path.to_string(),
-                runtime_route: host_domain::RuntimeRoute::LocalHttp {
-                    endpoint: "http://127.0.0.1:9999".to_string(),
-                },
-                started_at: started_at.to_string(),
-                descriptor: builtin_opencode_runtime_descriptor(),
-            },
-            &RuntimeStartupProgress {
-                started_at_instant,
-                started_at: started_at.to_string(),
-                attempts: Some(2),
-                elapsed_ms: Some(1000),
-            },
-        )?;
-
-        let status = service.runtime_startup_status("opencode", repo_path)?;
-
-        assert_eq!(status.stage, RepoRuntimeStartupStage::Idle);
-        assert!(status.runtime.is_none());
         Ok(())
     }
